LCP MOBILE (#4)
* chore: Update site title and font styles

* chore: Update site title, font styles, and dependencies

* refactor: Remove unused button element from README.md

* Remove unused RSS feed dependency and update site URL

* feat: Update Guide component with new images

The Guide component has been updated to use new images for the hand and text guide. This improves the visual experience for users.

* chore: Update background image URL to use webp format

The background image URL in the global CSS file has been updated to use the webp format instead of the png format. This change improves the loading time and performance of the website.