CONVERT WEBP (#6)
* chore: Update site title and font styles

* chore: Update site title, font styles, and dependencies

* refactor: Remove unused button element from README.md

* Remove unused RSS feed dependency and update site URL

* feat: Update Guide component with new images

The Guide component has been updated to use new images for the hand and text guide. This improves the visual experience for users.

* chore: Update background image URL to use webp format

The background image URL in the global CSS file has been updated to use the webp format instead of the png format. This change improves the loading time and performance of the website.

* refactor: Improve animation performance

Refactor the `animate` function in `animate.js` to improve animation performance. Instead of using `setTimeout`, use `requestAnimationFrame` for smoother animations. This change optimizes the animation rendering process and enhances the user experience.

* refactor: Update image URLs to use webp format